requirements.txt
django==1.9
python-dateutil==2.6.0
xlwt==1.2